requirements.txt
asgiref==3.11.1
crispy-tailwind==1.0.3
Django==5.0.14
django-admin-sortable2==2.3.1
django-ckeditor==6.7.3
django-crispy-forms==2.5
django-environ==0.12.1
django-extensions==4.1
django-js-asset==3.1.2
easy-thumbnails==2.10.1
pillow==12.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
sqlparse==0.5.5
text-unidecode==1.3
